move scienceBeamApi to xpub-client folder
