+++ server/xpub-controller/index.js
@@ -1,11 +1,11 @@
 const SupportingFiles = require('./supportingFiles')
 const Manuscript = require('./manuscript')
 const S3Storage = require('./client/s3Storage')
-const ScienceBeamApi = require('./client/scienceBeamApi')
+const ScienceBeamApi = require('../xpub-client/scienceBeamApi')
 
 module.exports = {
   SupportingFiles,
   Manuscript,
   S3Storage,
-  ScienceBeamApi
+  ScienceBeamApi,
 }
